requirements.txt
google-generativeai
python-dotenv
PyMuPDF
markdown
markdownify
streamlit
pypandoc
spacy
gensim
streamlit_pdf_viewer
streamlit==1.27.2
plotly